Extract serve.rs into http-service-hyper crate
